Move Fields class into its own file
